worker: build the per-uid agent on Linux
Unifying the two build paths needs the agent daemon on both
platforms. Factor the platform pieces of the agent (socket
activation, peer credentials, builder confinement, uid sweep) into
agent/darwin.rs and agent/linux.rs and enable the agent subcommand on
Linux, where builds still run unconfined.
